+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- `{BTree|Hash}Map.toHuman()` now display `Vec<u8>` keys as ascii (if detected)
+
+
 ## 8.13.1 Jul 16, 2022
 
 Changes: